+++ app/Http/Controllers/MyControllersController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\MyController;
+use Illuminate\Support\Facades\DB;
 
 class Data{
     public $id;
@@ -140,8 +141,22 @@ class MyControllersController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function destroy(MyController $controller)
+    public function destroy($id)
     {
+
+        $isExist = DB::table('kits')->where('controller_id', $id)->exists();
+        if($isExist)
+        {
+            return redirect('/'.$this->route)
+            ->with('warning', 'Bu '.$this->title.' türü diğer tablolarla ilişki olduğu için silemezsiniz.');
+        }       
+
+
+            MyController::find($id)->delete();
+            return redirect('/'.$this->route)
+                ->with('success', $this->title.' silme işlemi başarılı bir şekilde gerçekleştirildi');
+
+                /*
         try {
             $controller->delete();
         } catch (\Throwable $th) {
@@ -149,6 +164,6 @@ class MyControllersController extends Controller
             // mesaj gönderilecek.
             return redirect()->route($this->route.'.index');
         }
-        return redirect()->route($this->route.'.index');
+        return redirect()->route($this->route.'.index');*/
     }
 }
