+++ app/Http/Controllers/TypeController.php
@@ -6,6 +6,7 @@ use Illuminate\Http\Request;
 
 use App\Models\Type;
 use App\Models\Category;
+use DB;
 
 class Data{
     public $id;
@@ -140,9 +141,24 @@ class TypeController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function destroy(Type $type)
+    public function destroy($id)
     {
+        
+        $isExistPlant = DB::table('plants')->where('type_id', $id)->exists();
+        $isExistUnit = DB::table('units')->where('type_id', $id)->exists();
+        $isExistArea = DB::table('areas')->where('type_id', $id)->exists();
+        if($isExistPlant | $isExistUnit | $isExistArea)
+        {
+            return redirect('/'.$this->route)
+            ->with('warning', 'Bu '.$this->title.' türü diğer tablolarla ilişki olduğu için silemezsiniz.');
+        }       
+
+
+            Type::find($id)->delete();
+            return redirect('/'.$this->route)
+                ->with('success', $this->title.' silme işlemi başarılı bir şekilde gerçekleştirildi');
+        /*
         $type->delete();
-        return redirect()->route($this->route.'.index');
+        return redirect()->route($this->route.'.index');*/
     }
 }
